# dynamic wavelet matrix