--- a/.gitignore
+++ b/.gitignore
@@ -4,3 +4,4 @@ CMakeLists.txt~
 debug.cpp~
 gap_bv_benchmark.txt
 README.md~
+*~