This pulls in @MitI-7's [DynamicWaveletMatrix](https://github.com/MitI-7/WaveletMatrix/blob/master/DynamicWaveletMatrix/DynamicWaveletMatrix.hpp), rebases it on DYNAMIC's succinct bitvector, and partially matches the interface of wt_string.

It provides a fixed-width alphabet encoding requiring log(sigma) * N bits to encode N integers. I attempted to use it as a basis for the dynamic FM-index, but ran into problems that I assume relate to the fact that I'm not using an alphabet encoder like that in wt_string. However, I do pass the tests that @MitI-7 developed for the original implementation (these are in wm_string.cpp).

How can we improve this?