@ekg 2020-05-30
It's faster for access and most operations (except insert) than the original implementation:

```
build/wm_string | head -7
```

original:

```
SPEED
access:1357ms
rank:1551ms
select:2014ms
insert:1407ms
erase:2448ms
update:4574ms
```

This PR:

```
SPEED         
access:261ms  
rank:341ms    
select:1205ms 
insert:1534ms 
erase:1945ms  
update:4035ms 
```