@nicolaprezza 2020-05-30
thanks!! the next natural step would be to implement 2D range counting  and range reporting to turn this into a dynamic 2D range search data structure.