@ekg 2020-05-31
Out of curiosity, I'd also like to make it interchangeable with the
wt_string. The alphabet encoding issue is a curiosity. As would be getting
the matrix to be Huffman shaped.

On Sat, May 30, 2020, 19:43 Nicola Prezza <notifications@github.com> wrote:

> Merged #21 <https://github.com/xxsds/DYNAMIC/pull/21> into master.
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/xxsds/DYNAMIC/pull/21#event-3389925203>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AABDQENLOVM2YIJFYTLLSZLRUFASVANCNFSM4NOYR5SQ>
> .
>