@nicolaprezza 2020-06-01
Using other encodings with wavelet matrices is possible but trickier than wavelet trees: not any prefix-free code is good because in the matrix layout you don't want 'holes' to appear inside the bitvectors. What you need is to move the holes on the right as follows: if you want to assign k codes of length L (the lengths are those of Huffman, for example), then you sort all available binary codes of length L (i.e. those not having as prefix a shorter code) in decreasing co-lexicographic order, and pick the first k. This is described in chapter 6.2.5, paragraph "Compression" of Gonzalo's book "Compact Data Structures: A Practical Approach".

Unfortunately, Elias gamma and delta do not satisfy the above constraint, therefore those codes are not portable to a wavelet matrix. However, one could modify the Huffman alphabet encoder to return the codes described above, so that wm_string and wm_matrix are interchangeable at least on Huffman an fixed-length.